+++ modules/loaders.py
@@ -74,7 +74,6 @@ def get_all_params():
 
 
 def make_loader_params_visible(loader):
-    loader = fix_loader_name(loader)
     params = loaders_and_params[loader]
     all_params = get_all_params()
 
@@ -83,15 +82,3 @@ def make_loader_params_visible(loader):
         params += get_gpu_memory_keys()
 
     return [gr.update(visible=True) if k in params else gr.update(visible=False) for k in all_params]
-
-
-def fix_loader_name(name):
-    name = name.lower()
-    if name in ['llamacpp', 'llama.cpp', 'llama-cpp', 'llama cpp']:
-        return 'llama.cpp'
-    elif name in ['transformers', 'huggingface', 'hf', 'hugging_face', 'hugging face']:
-        return 'Transformers'
-    elif name in ['autogptq', 'auto-gptq', 'auto_gptq', 'auto gptq']:
-        return 'AutoGPTQ'
-    elif name in ['gptq-for-llama', 'gptqforllama', 'gptqllama', 'gptq for llama', 'gptq_for_llama']:
-        return 'GPTQ-for-LLaMa'
